% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_backup.R
\name{set_backup}
\alias{set_backup}
\title{A function for setting the core backup parameters for your library}
\usage{
set_backup(output)
}
\arguments{
\item{output}{Do you want to automatically save the output when backing up the library? Default to TRUE}
}
\description{
This function allows you to set
}
\examples{
set_backup()

}
\keyword{output}
